Move all template RCP classes to new package
