+++ sernet.gs.ui.rcp.main/src/sernet/verinice/rcp/templates/TemplateSelectionDialog.java
@@ -17,7 +17,7 @@
  * Contributors:  
  *     Viktor Schmidt <vschmidt[at]ckc[dot]de> - initial API and implementation  
  ******************************************************************************/
-package sernet.verinice.rcp;
+package sernet.verinice.rcp.templates;
 
 import org.eclipse.swt.graphics.Point;
 import org.eclipse.swt.widgets.Composite;
